requirements.txt
PyQT5==5.15.10
PyOpenGL==3.1.7
numpy==1.26.1
matplotlib==3.8.0
jsonschema==4.20.0